+++ crates/mdk-sqlite-storage/src/groups.rs
@@ -386,6 +386,53 @@ impl GroupStorage for MdkSqliteStorage {
         })
     }
 
+    fn get_group_legacy_exporter_secret(
+        &self,
+        mls_group_id: &GroupId,
+        epoch: u64,
+    ) -> Result<Option<GroupExporterSecret>, GroupError> {
+        if self.find_group_by_mls_group_id(mls_group_id)?.is_none() {
+            return Err(GroupError::InvalidParameters("Group not found".to_string()));
+        }
+
+        self.with_connection(|conn| {
+            let mut stmt = conn
+                .prepare(
+                    "SELECT * FROM group_exporter_secrets WHERE mls_group_id = ? AND epoch = ? AND label = 'legacy-group-event'",
+                )
+                .map_err(into_group_err)?;
+
+            stmt.query_row(
+                params![mls_group_id.as_slice(), epoch],
+                db::row_to_group_exporter_secret,
+            )
+            .optional()
+            .map_err(into_group_err)
+        })
+    }
+
+    fn save_group_legacy_exporter_secret(
+        &self,
+        group_exporter_secret: GroupExporterSecret,
+    ) -> Result<(), GroupError> {
+        if self
+            .find_group_by_mls_group_id(&group_exporter_secret.mls_group_id)?
+            .is_none()
+        {
+            return Err(GroupError::InvalidParameters("Group not found".to_string()));
+        }
+
+        self.with_connection(|conn| {
+            conn.execute(
+                "INSERT OR REPLACE INTO group_exporter_secrets (mls_group_id, epoch, secret, label) VALUES (?, ?, ?, 'legacy-group-event')",
+                params![&group_exporter_secret.mls_group_id.as_slice(), &group_exporter_secret.epoch, group_exporter_secret.secret.as_ref()],
+            )
+            .map_err(into_group_err)?;
+
+            Ok(())
+        })
+    }
+
     fn get_group_mip04_exporter_secret(
         &self,
         mls_group_id: &GroupId,